split node and controller
