+++ lbr_demos/lbr_demos_fri_ros2_advanced_python/lbr_demos_fri_ros2_advanced_python/admittance_rcm_controller.py
@@ -1,28 +1,8 @@
-#!/usr/bin/python3
-import time
-
 import casadi as cs
-import numpy as np
-import optas  # https://github.com/cmower/optas
-import rclpy
-from lbr_fri_msgs.msg import LBRPositionCommand, LBRState
-from rclpy.node import Node
+import optas
 
 
-def dampedLeastSquares(J, lam=2e-1):
-    U, S, VT = np.linalg.svd(J)
-
-    # print(f"{U.shape=}, {S.shape=}, {VT.shape=}")
-    ns = S.shape[0]
-    s = np.zeros((J.shape[1], J.shape[0]))
-    for i in range(ns):
-        s[i, i] = S[i] / (S[i] * S[i] + lam * lam)
-    # print("s=", s)
-    return VT.T @ s @ U.T
-
-
-class Controller:
-
+class AdmittanceRCMController:
     def __init__(self, robot_description):
         self._solver_dur = None
         ee_link = "link_ee"
@@ -57,9 +37,6 @@ class Controller:
 
         builder.add_equality_constraint("rcm", self.dist_sqr(qF, rcm))
 
-        xlim = -0.15
-        zlim = 0.8
-
         c1 = cs.sumsqr(qd - qd_goal)
         builder.add_cost_term("match_qd_goal", c1)
 
@@ -67,7 +44,6 @@ class Controller:
         builder.add_equality_constraint("dynamics", q0 + qd, qF)
 
         self.eff_transform = robot.get_global_link_transform_function(ee_link)
-        pF = robot.get_global_link_position(ee_link, qF)
 
         q = cs.SX.sym("q", robot.ndof)
         p = robot.get_global_link_position(ee_link, q)
@@ -99,111 +75,3 @@ class Controller:
 
     def get_qd_target(self):
         return self.sol[f"{self.name}/dq"].toarray().flatten()
-
-
-class LBRAdmittanceControlRCMNode(Node):
-
-    def __init__(self, node_name="admittance_rcm_control_node"):
-        super().__init__(node_name)
-
-        # declare and get parameters
-        self.declare_parameter("command_rate", 100)
-        self.declare_parameter("robot_description")
-
-        self._command_rate = int(self.get_parameter("command_rate").value)
-        self.dt = 1.0 / float(self._command_rate)
-        self._robot_description = str(self.get_parameter("robot_description").value)
-
-        self.dq = np.zeros(7)
-        self._controller = Controller(self._robot_description)
-
-        self._lbr_state_subscriber = self.create_subscription(
-            LBRState, "/lbr/state", self._lbr_state_callback, 1
-        )
-
-        self._lbr_command_publisher = self.create_publisher(
-            LBRPositionCommand, "/lbr/command/joint_position", 1
-        )
-
-    def command(self, q):
-        lbr_command = LBRPositionCommand()
-        lbr_command.joint_position = q
-        self._lbr_command_publisher.publish(lbr_command)
-
-    def _admittance(self, tau_ext, qc):
-
-        J = self._controller.Jac(qc).toarray()
-
-        Jinv = dampedLeastSquares(J)
-        f_ext = Jinv.T @ tau_ext
-
-        dx = np.zeros(6)
-
-        translational_vel = 0.12
-        th_f = 4.0
-        th_tau = 1.0
-        scale = 2.0
-
-        for i in [0, 1, 2]:
-            sign = np.sign(f_ext[i])
-            if (abs(f_ext[i]) > th_f) and (abs(f_ext[i]) < scale * th_f):
-                dx[i] = (
-                    sign
-                    * translational_vel
-                    * (abs(f_ext[i]) - th_f)
-                    / (scale * th_f - th_f)
-                )
-            elif abs(f_ext[i]) > scale * th_f:
-                if f_ext[i] > 0.0:
-                    dx[i] = translational_vel
-                else:
-                    dx[i] = -translational_vel
-
-        for i in [3, 4, 5]:
-            sign = np.sign(f_ext[i])
-            if (abs(f_ext[i]) > th_tau) and (abs(f_ext[i]) < scale * th_tau):
-                dx[i] = (
-                    sign
-                    * translational_vel
-                    * (abs(f_ext[i]) - th_tau)
-                    / (scale * th_tau - th_tau)
-                )
-            elif abs(f_ext[i]) > scale * th_tau:
-                if f_ext[i] > 0.0:
-                    dx[i] = translational_vel
-                else:
-                    dx[i] = -translational_vel
-
-        dq = Jinv @ dx
-
-        self.dq = (1 - 0.02) * self.dq + 0.02 * dq
-
-        return self.dq
-
-    def _lbr_state_callback(self, msg: LBRState) -> None:
-
-        if self._controller._rcm is None:
-            self._controller.set_start(msg.measured_joint_position)
-            return
-
-        dq_goal = self._admittance(msg.external_torque, msg.measured_joint_position)
-        self._controller.reset(msg.measured_joint_position, dq_goal)
-
-        if self._controller.solve():
-            qc = np.array(msg.measured_joint_position)
-            dq = self._controller.get_qd_target()
-            qn = qc + self.dt * dq * 10.0
-            self.command(qn.tolist())
-        else:
-            self.get_logger().error("Solver failed!")
-
-
-def main(args=None):
-    rclpy.init(args=args)
-    node = LBRAdmittanceControlRCMNode()
-    rclpy.spin(node)
-    rclpy.shutdown()
-
-
-if __name__ == "__main__":
-    main()
